+++ src/onyx/triggers/refinements.clj
@@ -1,11 +1,15 @@
 (ns onyx.triggers.refinements)
 
 (def discarding 
-  {:refinement/state-update (fn [event trigger opts state])
-   :refinement/apply-state-update (fn [event trigger state entry]
+  {:refinement/create-state-update (fn discarding-create-state-update 
+                                     [trigger opts state])
+   :refinement/apply-state-update (fn discarding-apply-state-update 
+                                    [trigger opts state entry]
                                     {})})
 
 (def accumulating
-  {:refinement/state-update (fn [event trigger opts state])
-   :refinement/apply-state-update (fn [event trigger state entry]
+  {:refinement/create-state-update (fn accumulating-create-state-update 
+                                     [trigger opts state])
+   :refinement/apply-state-update (fn accumulating-apply-state-update
+                                    [trigger opts state entry]
                                     state)})