+++ test/onyx/windowing/aggregation_conj_test.clj
@@ -70,7 +70,7 @@
 
 (def test-state (atom []))
 
-(defn update-atom! [event window trigger {:keys [window extent->bounds changelog] :as opts} state]
+(defn update-atom! [event window trigger {:keys [window/extent->bounds] :as opts} state]
   (doall 
     (map (fn [[extent extent-state] [lower-bound upper-bound]] 
            (swap! test-state conj [lower-bound upper-bound extent-state]))